CentralWidget: Replaced pykdeuic4 generated code with cleaned up original ui file. - Merged with UI_CentralWidget and moved to centralwidget.py - Moved the QTreeWidget that displays the directory structure  with its helper classes to autokey_treewidget.py - Added some application logging to CentralWidget and ConfigWindow classes. - Some PEP8 variable naming fixes and other minor style issues.
