+++ lib/autokey/qtui/resources/resources.qrc
@@ -9,5 +9,6 @@
     <file>ui/scriptpage.ui</file>
     <file>ui/phrasepage.ui</file>
     <file>ui/folderpage.ui</file>
+    <file>ui/centralwidget.ui</file>
 </qresource>
 </RCC>
